Organized kotlin extension functions code in apt. packages
